Refactor server package.
